+++ include/gtest/switch/iutest_switch_gmock.hpp
@@ -130,6 +130,16 @@ namespace matchers
     template<typename T>
     inline internal::EqMatcher<T> Equals(T x) { return Eq(x); }
 
+    template<typename T>
+    inline internal::FloatingEqMatcher<T> FloatingPointEq(T rhs) {
+        return internal::FloatingEqMatcher<T>(rhs, false);
+    }
+
+    template<typename T>
+    inline internal::FloatingEqMatcher<T> NanSensitiveFloatingPointEq(T rhs) {
+        return internal::FloatingEqMatcher<T>(rhs, true);
+    }
+
 #if GMOCK_VER >= 0x01070000
     // gmock 1.7 later
     using ::testing::IsEmpty;