+++ include/iutest_ver.hpp
@@ -17,11 +17,11 @@
 
 //======================================================================
 // define
-#define IUTEST_VER          0x01160502u //!< iutest version 1.16.5.2
+#define IUTEST_VER          0x01160504u //!< iutest version 1.16.5.4
 #define IUTEST_MAJORVER     0x01u       //!< Major Version
 #define IUTEST_MINORVER     0x16u       //!< Minor Version
 #define IUTEST_MICROVER     0x05u       //!< Micro Version
-#define IUTEST_REVISION     0x02u       //!< Revision
+#define IUTEST_REVISION     0x04u       //!< Revision
 
 #define IUTEST_BUILD        IUTEST_MICROVER //!< @deprecated
 
@@ -120,9 +120,12 @@
  * @page    CHANGELOG   変更履歴
  * @par     Changes
  <ul>
-  <li>v1.16.5.0
+  <li>v1.16.6.0
     <ul>
-      <li></li>
+      <li>long double に対応</li>
+      <li>iuwandbox: -D__WANDBOX__ コンパイラーオプションを必ずつけるように変更</li>
+      <li>iuwandbox: dryrun オプションが失敗を返す問題を修正</li>
+      <li>Fixed: いくつかのコンパイラー警告を修正</li>
     </ul>
   </li>
   <li>v1.16.5.0
@@ -132,7 +135,6 @@
       <li>ad_hoc_testresult の取得関数名を ad_hoc_test_result に変更</li>
       <li>compatibility: iuutil に ad_hoc_testresult の取得関数を追加</li>
       <li>iuwandbox で std-cxx の指定をするとオプションチェックで失敗する不具合を修正</li>
-      <li></li>
     </ul>
   </li>
   <li>v1.16.4.0